llm_roi_study/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
pydantic==2.7.1
httpx==0.27.0
faiss-cpu==1.13.2
numpy==1.26.4
ragas==0.1.9
datasets==2.19.1
tiktoken==0.7.0
pandas==2.2.2
scipy==1.13.0
statsmodels==0.14.2
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
pyyaml==6.0.1
python-dotenv==1.0.1
tqdm==4.66.4
pytest==8.2.0
